CPL/requirements.txt
torch==2.0.1
transformers==4.10.0
numpy==1.24.4
scikit-learn==0.24.2
openai==0.28.0
nltk==3.8.1
retry==0.9.2
google-generativeai==0.5.2